refactor(useCouponList.ts): hook의 이름을 useAvailableCouponList에서 useCouponList로 변경
